+++ src/App.tsx
@@ -18,8 +18,8 @@ import FindPasswordPage from 'pages/Auth/FindPasswordPage';
 import StoreDetailPage from 'pages/Store/StoreDetailPage';
 import StoreBenefitPage from 'pages/Store/StoreBenefitPage';
 import CampusInfo from 'pages/CampusInfo';
-import BusRoutePage from 'pages/BusRoutePage';
-import BusCoursePage from 'pages/BusCoursePage';
+import BusRoutePage from 'pages/Bus/BusRoutePage';
+import BusCoursePage from 'pages/Bus/BusCoursePage';
 import IndexPage from 'pages/IndexPage';
 import RoomPage from 'pages/Room/RoomPage';
 import RoomDetailPage from 'pages/Room/RoomDetailPage';
