+++ src/pages/Bus/BusCoursePage/BusTimetableDetail/index.tsx
@@ -1,9 +1,9 @@
-import useShuttleTimetableDetail from 'pages/BusCoursePage/hooks/useShuttleTimetableDetail';
+import useShuttleTimetableDetail from 'pages/Bus/BusCoursePage/hooks/useShuttleTimetableDetail';
 import { useEffect, useState } from 'react';
 import BusIcon from 'assets/svg/bus.svg';
 import { cn } from '@bcsdlab/utils';
 import useMediaQuery from 'utils/hooks/layout/useMediaQuery';
-import InfomationIcon from 'assets/svg/info-gray.svg';
+import InfomationIcon from 'assets/svg/Bus/info-gray.svg';
 import styles from './BusTimetableDetail.module.scss';
 
 interface ShuttleTimetableDetailProps {
