+++ src/pages/Bus/BusRoutePage/components/TimeDetail/TimeDetail.module.scss
